inspektor: Rename reindent to indent
